refactorin to i/o and expand functionality
